@@ -1,26 +1,23 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
 import Vue from 'vue'
-import App from './App'
+import App from './App.vue'
 import router from './router'
 import store from './store'
-import http from './common/js/http'
-// vux 引入组件
-import { AlertPlugin, ToastPlugin } from 'vux'
-Vue.use(AlertPlugin)
-Vue.use(ToastPlugin)
-// axios封装
-Vue.use(http)
-Vue.config.productionTip = false
-// 解决点击延迟
-const FastClick = require('fastclick')
-FastClick.attach(document.body)
+// 导入所有组件
+import Vant from 'vant'
+import 'vant/lib/index.css'
 
-/* eslint-disable no-new */
+import api from '@/api'
+
+import './styles/index.scss' /* 引入样式 */
+
+Vue.use(Vant)
+// api
+Vue.prototype.$api = api
+
+Vue.config.productionTip = false
+
 new Vue({
-    el: '#app',
-    router,
-    store,
-    template: '<App/>',
-    components: { App }
-})
+  router,
+  store,
+  render: h => h(App)
+}).$mount('#app')
