@@ -1,34 +1,33 @@
-// 加入购物车中的商品
-export const cartData = state => state.cartData
-// 计算总数量
-export const totalCount = (state, getters) => {
+const getters = {
+  userInfo: state => state.user.userInfo,
+  cartList: state => state.cart.cartList,
+  checkedGoodsList: state => {
+    let tempArr = []
+    state.cart.cartList.forEach(item => {
+      if (item.isChecked) {
+        tempArr.push(item)
+      }
+    })
+    return tempArr
+  },
+  totalCount: state => {
     let num = 0
-    getters.cartData.forEach(item => {
-        num += item.count
+    state.cart.cartList.forEach(item => {
+      num += item.count
     })
     return num
-}
-// 计算购物车选中的总数量
-export const selectedArr = (state, getters) => {
-    let selectArr = []
-    getters.cartData.forEach(item => {
-        if (item.isChecked) {
-            selectArr.push(item)
-        }
-    })
-    return selectArr
-}
-// 计算总价
-export const tolalPrice = (state, getters) => {
-    // 避免 0.05 + 0.01 = 0.060 000 000 000 000 005 的问题，乘以 100 *100
+  },
+  tolalPrice: state => {
+    // 避免 0.05 + 0.01 = 0.060 000 000 000 000 005 的问题，=> 0.05x100 * 0.01 x100 / 100 * 100
     const multiple = 100
-    let total = 0
-    getters.cartData.forEach(item => {
-        if (item.isChecked) {
-            total += item.price * multiple * item.count * multiple
-        }
+    let amount = 0
+    state.cart.cartList.forEach(item => {
+      if (item.isChecked) {
+        amount += item.price * multiple * item.count * multiple
+      }
     })
-    return (total = total / (multiple * multiple))
+    return (amount = amount / (multiple * multiple))
+  }
 }
 
-export const userInfo = state => state.userInfo
+export default getters
