@@ -1,25 +1,28 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-
-import * as actions from './actions'
-import * as getters from './getters'
-import mutations from './mutations'
-import state from './state'
+import app from './modules/app'
+import user from './modules/user'
+import cart from './modules/cart'
+import getters from './getters'
 
 // 开发模式下使用日志，和谷歌浏览器里面的vuedevtool插件类似
 // 子修改state的时候会在控制台打印一些信息
 import createLogger from 'vuex/dist/logger'
 
-Vue.use(Vuex)
-
 // 调试,开发模式开启严格模式，要使用mutations提交
 const debug = process.env.NODE_ENV !== 'production'
 
-export default new Vuex.Store({
-    actions,
-    getters,
-    state,
-    mutations,
-    strict: debug,
-    plugins: debug ? [createLogger()] : []
+Vue.use(Vuex)
+
+const store = new Vuex.Store({
+  modules: {
+    app,
+    user,
+    cart
+  },
+  getters,
+  strict: debug,
+  plugins: debug ? [createLogger()] : []
 })
+
+export default store
